+++ globals/fields.go
@@ -5,6 +5,7 @@ package globals
 
 const (
 	IdField                                     = "id"
+	IdsField                                    = "ids"
 	VersionField                                = "version"
 	NameField                                   = "name"
 	DescriptionField                            = "description"