+++ CHANGELOG.md
@@ -6,6 +6,21 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 
 ### New and Improved
 
+* roles: In grants, the `id` field has been changed to `ids` (but `id` will
+  still be accepted for now, up until 0.15.0). In the `ids` field, multiple IDs
+  can now be specified in a grant, either via commas (text format) or array
+  (JSON format). ([PR](https://github.com/hashicorp/boundary/pull/3263)).
+
+### Deprecations/Changes
+
+* Grants can now accept more than one ID per grant string (or entry in JSON) via
+  the `ids` parameter. In 0.15.0 the ability to add new grants via the `id`
+  parameter will be removed.
+
+## 0.13.0
+
+### New and Improved
+
 * KMS workers: KMS workers now have feature parity with PKI workers (they
   support multi-hop and Vault private access) and support separate KMSes for
   authenticating downstreams across different networks. See the [worker